+++ CI/SCRIPTS/test_satellites.py
@@ -144,11 +144,16 @@ def _test_core(
             LOGGER.info("Checking opening solutions")
             LOGGER.info("MTD")
             prod: Product = READER.open(path, method=CheckMethod.MTD, remove_tmp=False)
-            LOGGER.info("NAME")
-            prod_name = READER.open(path, method=CheckMethod.NAME)
-            LOGGER.info("BOTH")
-            prod_both = READER.open(path, method=CheckMethod.BOTH)
             assert prod is not None
+
+            LOGGER.info("NAME")
+            prod_name = READER.open(
+                path, method=CheckMethod.NAME, constellation=prod.constellation
+            )
+            LOGGER.info("BOTH")
+            prod_both = READER.open(
+                path, method=CheckMethod.BOTH, constellation=prod.constellation
+            )
             assert prod_name is not None
             assert prod_both is not None
             assert prod == prod_name